+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - whisper.cpp library for speech recognition
 - English speech recognition, with the result logged
 - Transcript data structure
-- Record transcript
+- Add transcriptions to audio entries
+- Capture duration of generating transcript
 
 ## [0.8.344] - 2023-05-05
 ### Changed: